+++ app.py
@@ -2,7 +2,8 @@ import logging
 
 import dash
 
-from back import MongoManager
+from data import MongoManager
+from logic import AppLogic
 
 logger = logging.getLogger('dashvision.{}'.format(__name__))
 
@@ -12,4 +13,5 @@ app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
 app.config.suppress_callback_exceptions = True
 
 data_manager = MongoManager(server='drunk', port='27017')
+logic_manager = AppLogic(data_manager)
 default_columns = ['_id', 'status']
