+++ apps/datatable.py
@@ -30,4 +30,4 @@ layout = html.Div([
 
 @app.callback(Output('selected_datatable', 'data'), [Input('tab-data', 'data')])
 def populate_table(data_args):
-    return data_manager.get_table_content_from_ids(data_args['db'], data_args['selected_ids'], default_columns)
+    return data_manager._get_table_content_from_ids(data_args['db'], data_args['selected_ids'], default_columns)
