+++ index.py
@@ -7,7 +7,7 @@ import time
 from dash.dependencies import Input, Output, State
 from dash.exceptions import PreventUpdate
 
-from app import app, default_columns, data_manager
+from app import app, default_columns, logic_manager
 from apps import config_viewer, datatable
 
 logging.basicConfig(level=logging.DEBUG)
@@ -36,7 +36,7 @@ app.layout = html.Div([
     html.Div([
         html.Div('Select DataBase'),
         html.Div(dcc.Dropdown(id='database-selector',
-                              options=data_manager.on_load_database_options()))
+                              options=logic_manager.on_load_database_options()))
     ]),
 
     # Select experiments
@@ -110,7 +110,7 @@ def select_db(ts, db_options, stored_db_name):
     if stored_db_name is None or stored_db_name not in all_db_names:
         raise PreventUpdate
 
-    exp_names = data_manager.get_experiment_names(stored_db_name)
+    exp_names = logic_manager.get_experiment_names(stored_db_name)
     exps_selector_options = ([{'label': name, 'value': name} for name in exp_names])
     return stored_db_name, exps_selector_options
 
@@ -173,10 +173,10 @@ def update_experiment_table(experiment_names, cols, query_nsub, db_name, query):
         return []
 
     columns = [col['id'] for col in cols]
-    rows = data_manager.get_table_content_from_exp_names(db_name, experiment_names, columns)
+    rows = logic_manager.get_table_content_from_exp_names(db_name, experiment_names, columns)
 
     if query_nsub is not None:
-        rows = data_manager.filter_rows_by_query(rows, query)
+        rows = logic_manager.filter_rows_by_query(rows, query)
 
     return rows
